standalone extensions lib
