+++ lib/extensions/src/lib/components/dynamic-component/dynamic.component.ts
@@ -22,6 +22,7 @@ import { ExtensionComponent } from '../../services/component-register.service';
 // cSpell:words lifecycle
 @Component({
     selector: 'adf-dynamic-component',
+    standalone: true,
     template: `<div #content></div>`
 })
 export class DynamicExtensionComponent implements OnChanges, OnDestroy {
